+++ aggregates/aaBoundingBox.h
@@ -6,13 +6,13 @@ class AABoundingBox : public GeometricObject
 {
 	public:
 		AABoundingBox(GeometricObject *object);
-        virtual void getBounds(double min[3], double max[3]) const;
+        virtual void getBounds(float min[3], float max[3]) const;
 		virtual ShadeRec hit(Ray &ray) const;
 		bool testBBox(Ray &ray) const;
 	private:
 		GeometricObject *m_object;
-		double m_min[3];
-		double m_max[3];
+		float m_min[3];
+		float m_max[3];
 };
 
 #endif
