+++ aggregates/collection.h
@@ -8,12 +8,12 @@ class Collection : public GeometricObject
 	public:
 		Collection();
 		void addObject(GeometricObject *object);
-        virtual void getBounds(double min[3], double max[3]) const;
+        virtual void getBounds(float min[3], float max[3]) const;
 		virtual ShadeRec hit(Ray &ray) const;
 	private:
 		std::vector<GeometricObject*> m_objects;
-		double m_min[3];
-		double m_max[3];
+		float m_min[3];
+		float m_max[3];
 };
 
 #endif
