+++ brdf/lambert.cpp
@@ -1,9 +1,9 @@
 #include "lambert.h"
 
-double Lambert::shade(ShadeRec &shadeRec, Vec3 &lightDir) const
+float Lambert::shade(ShadeRec &shadeRec, Vec3 &lightDir) const
 {
 	Vec3 normalVec = shadeRec.getNormal().toVec3();
-	double val = lightDir.dot(normalVec);
+	float val = lightDir.dot(normalVec);
 	if(val < 0.0){
 		val = 0.0;
 	}
