+++ brdf/lambert.h
@@ -6,7 +6,7 @@ class Lambert: public BRDF
 {
 	public:
 		Lambert(){};
-		virtual double shade(ShadeRec &shadeRec, Vec3 &lightDir) const;
+		virtual float shade(ShadeRec &shadeRec, Vec3 &lightDir) const;
 	private:
 };
