+++ geom/plane.cpp
@@ -1,7 +1,7 @@
 #include "plane.h"
 #include <cfloat>
 
-void Plane::getBounds(double min[3], double max[3]) const
+void Plane::getBounds(float min[3], float max[3]) const
 {
     min[0] = -0.1;
     max[0] = 0.1;
@@ -14,12 +14,12 @@ void Plane::getBounds(double min[3], double max[3]) const
 ShadeRec Plane::hit(Ray &ray) const
 {
 	ShadeRec sr;
-    double origin_z = ray.m_origin[2];
-    double direction_z = ray.m_dir[2];
+    float origin_z = ray.m_origin[2];
+    float direction_z = ray.m_dir[2];
     if(direction_z == 0.0){
         return sr;
     }
-    double t = -origin_z/direction_z;
+    float t = -origin_z/direction_z;
     if(t>0.0){
         sr.setHit(true);
         sr.setHitT(t);
