+++ geom/sphere.cpp
@@ -1,7 +1,7 @@
 #include "sphere.h"
 #include "../utils/matrix4x4.h"
 
-void Sphere::getBounds(double min[3], double max[3]) const
+void Sphere::getBounds(float min[3], float max[3]) const
 {
     for (int i = 0; i < 3; i++) {
         min[i] = -m_radius;
@@ -13,28 +13,28 @@ ShadeRec Sphere::hit(Ray &ray) const
 {
 	Point pnt = ray.m_origin;
 	Vec3 o = pnt.toVec3();
-	double a = ray.m_dir.dot(ray.m_dir);
-	double b = 2.0*(ray.m_dir.dot(o));
-	double c = o.dot(o) - m_radiusSquared;
+	float a = ray.m_dir.dot(ray.m_dir);
+	float b = 2.0*(ray.m_dir.dot(o));
+	float c = o.dot(o) - m_radiusSquared;
 
 	ShadeRec sr;
 
-	double discriminant = b*b - 4*a*c;
+	float discriminant = b*b - 4*a*c;
 	if(discriminant < 0){
 		return sr;
 	}
-	double discSqrt = sqrt(discriminant);
-	double q;
+	float discSqrt = sqrt(discriminant);
+	float q;
 	if(b<0.0){
 		q = (-b - discSqrt)/2.0;
 	}
 	else{
 		q = (-b + discSqrt)/2.0;
 	}
-	double t0 = q/a;
-	double t1 = c/q;
+	float t0 = q/a;
+	float t1 = c/q;
 	if(t1<t0){
-		double tmp = t0;
+		float tmp = t0;
 		t0 = t1;
 		t1 = tmp;
 	}
