+++ lights/directional.cpp
@@ -1,6 +1,6 @@
 #include "directional.h"
 
-double Directional::computeStrength(Point pos)
+float Directional::computeStrength(Point pos)
 {
 	return 1.0;
 }
