+++ lights/point.h
@@ -5,8 +5,8 @@
 class PointLight: public Light
 {
 	public:
-		PointLight(double strength, Matrix4x4 transform):Light(strength,transform){};
-        virtual double computeStrength(Point pos);
+		PointLight(float strength, Matrix4x4 transform):Light(strength,transform){};
+        virtual float computeStrength(Point pos);
         virtual Vec3 computeDirection(Point pos);
 	private:
 };
