+++ utils/normal.h
@@ -7,8 +7,8 @@ class Normal
 	public:
 		Normal():m_vec(){};
 		Normal(Vec3 &vec):m_vec(vec){};
-		Normal(double x, double y, double z):m_vec(x,y,z){};
-		Normal(double d[3]):m_vec(d){};
+		Normal(float x, float y, float z):m_vec(x,y,z){};
+		Normal(float d[3]):m_vec(d){};
 		~Normal(){};
 		Vec3 toVec3(){return m_vec;};
 		void normalize(){m_vec.normalize();};
