+++ utils/point.h
@@ -7,14 +7,14 @@ class Point
 	public:
 		Point():m_vec(){};
 		Point(Vec3 &vec):m_vec(vec){};
-		Point(double x, double y, double z):m_vec(x,y,z){};
-		Point(double d[3]):m_vec(d){};
+		Point(float x, float y, float z):m_vec(x,y,z){};
+		Point(float d[3]):m_vec(d){};
 		~Point(){};
-		double x() {return m_vec.x();};
-		double y() {return m_vec.y();};
-		double z() {return m_vec.z();};
-		double& operator[](int id){assert(id < 3 && id>=0);return m_vec[id];};
-		double distance(Point &other){return (m_vec - other.m_vec).magnitude();};
+		float x() {return m_vec.x();};
+		float y() {return m_vec.y();};
+		float z() {return m_vec.z();};
+		float& operator[](int id){assert(id < 3 && id>=0);return m_vec[id];};
+		float distance(Point &other){return (m_vec - other.m_vec).magnitude();};
 		Vec3 operator-(Point &other);
 		Point& operator-=(Vec3 &vec);
 		Point operator-(Vec3 &vec);
