+++ utils/ray.h
@@ -6,16 +6,16 @@
 class Ray
 {
 	public:
-		Ray(double tMin=0.0, int depth =0):m_origin(),m_dir(),m_tMin(tMin),m_depth(depth){};
-		Ray(Point origin, Vec3 dir, double tMin=0.0, int depth=0):m_origin(origin),m_dir(dir),m_tMin(tMin),m_depth(depth){};
-		Ray(Vec3 dir, double tMin=0.0, int depth=0):m_origin(),m_dir(dir),m_tMin(tMin),m_depth(depth){};
-		Ray(Point origin, double tMin=0.0, int depth=0):m_origin(origin),m_dir(),m_tMin(tMin),m_depth(depth){};
+		Ray(float tMin=0.0, int depth =0):m_origin(),m_dir(),m_tMin(tMin),m_depth(depth){};
+		Ray(Point origin, Vec3 dir, float tMin=0.0, int depth=0):m_origin(origin),m_dir(dir),m_tMin(tMin),m_depth(depth){};
+		Ray(Vec3 dir, float tMin=0.0, int depth=0):m_origin(),m_dir(dir),m_tMin(tMin),m_depth(depth){};
+		Ray(Point origin, float tMin=0.0, int depth=0):m_origin(origin),m_dir(),m_tMin(tMin),m_depth(depth){};
 
-		Point getPointAtPos(double t);
+		Point getPointAtPos(float t);
 
 		friend std::ostream& operator<<(std::ostream &out, Ray &vec);
 
-		double m_tMin;
+		float m_tMin;
 		Point m_origin;
 		Vec3 m_dir;
 		int m_depth;
