+++ utils/vec3.cpp
@@ -1,13 +1,13 @@
 #include "vec3.h"
 
-double Vec3::dot(Vec3 &other)
+float Vec3::dot(Vec3 &other)
 {
 	return m_d[0]*other.m_d[0] + m_d[1]*other.m_d[1] + m_d[2]*other.m_d[2];
 }
 
 Vec3 Vec3::cross(Vec3 &rhs)
 {
-	double result[3];
+	float result[3];
 	result[0] = y()*rhs.z() - z()*rhs.y();
 	result[1] = z()*rhs.x() - x()*rhs.z();
 	result[2] = x()*rhs.y() - y()*rhs.x();
@@ -17,7 +17,7 @@ Vec3 Vec3::cross(Vec3 &rhs)
 Vec3 Vec3::reflect(Vec3 &other)
 {
 	Vec3 thisVec = Vec3(m_d);
-	double val = thisVec.dot(other);
+	float val = thisVec.dot(other);
 	val *= 2.0;
 	Vec3 result = thisVec*val - other;
 	return result;
@@ -25,9 +25,9 @@ Vec3 Vec3::reflect(Vec3 &other)
 
 void Vec3::normalize()
 {
-	double mag = magnitude();
+	float mag = magnitude();
 	if(mag != 0){
-		double invmag = 1.0/mag;
+		float invmag = 1.0/mag;
 		m_d[0] = m_d[0]*invmag;
 		m_d[1] = m_d[1]*invmag;
 		m_d[2] = m_d[2]*invmag;
@@ -78,7 +78,7 @@ Vec3 Vec3::operator-(Vec3 &other)
 	return tmp;
 }
 
-Vec3& Vec3::operator*=(double &other)
+Vec3& Vec3::operator*=(float &other)
 {
 	m_d[0] *= other;
 	m_d[1] *= other;
@@ -86,7 +86,7 @@ Vec3& Vec3::operator*=(double &other)
 	return *this;
 }
 
-Vec3 Vec3::operator*(double &other)
+Vec3 Vec3::operator*(float &other)
 {
 	Vec3 tmp = *this;
 	tmp *= other;
