+++ viewPlane/viewPlane.h
@@ -10,10 +10,10 @@
 class ViewPlane
 {
     public:
-        ViewPlane(int resX, int resY, double sizeX, double sizeY);
-        ViewPlane(int resX, int resY, double size[2]){ ViewPlane(resX,resY, size[0], size[1]);};
-        ViewPlane(int res[2], double size[2]){ ViewPlane(res[0],res[1], size[0], size[1]);};
-        ViewPlane(int res[2], double sizeX, double sizeY){ ViewPlane(res[0],res[1], sizeX, sizeY);};
+        ViewPlane(int resX, int resY, float sizeX, float sizeY);
+        ViewPlane(int resX, int resY, float size[2]){ ViewPlane(resX,resY, size[0], size[1]);};
+        ViewPlane(int res[2], float size[2]){ ViewPlane(res[0],res[1], size[0], size[1]);};
+        ViewPlane(int res[2], float sizeX, float sizeY){ ViewPlane(res[0],res[1], sizeX, sizeY);};
         ~ViewPlane(){};
         int getNumPixels(){return m_pixels.size();};
         Ray getPixelRay(int index, Point sample);
@@ -24,7 +24,7 @@ class ViewPlane
     private:
         int m_resolution[2];
         std::vector<RGBA> m_pixels;
-        double m_size[2];
+        float m_size[2];
 };
 
 #endif
